+++ lib/src/messages/languages/es_msg.dart
@@ -12,6 +12,12 @@ class EspanaMessages implements Messages {
   @override
   String suffixAgo() => '';
 
+  @override
+  String prefixFromNow() => 'en';
+
+  @override
+  String suffixFromNow() => '';
+
   /// Message when the elapsed time is less than 15 seconds.
   @override
   String justNow(int seconds) => 'hace poco';