+++ lib/src/messages/languages/fr_msg.dart
@@ -12,6 +12,12 @@ class FrenchMessages implements Messages {
   @override
   String suffixAgo() => '';
 
+  @override
+  String prefixFromNow() => 'dans';
+
+  @override
+  String suffixFromNow() => '';
+
   /// Message when the elapsed time is less than 15 seconds.
   @override
   String justNow(int seconds) => 'en ce moment';