+++ lib/src/messages/languages/oc_msg.dart
@@ -12,6 +12,12 @@ class OccitanMessages implements Messages {
   @override
   String suffixAgo() => '';
 
+  @override
+  String prefixFromNow() => 'dins';
+
+  @override
+  String suffixFromNow() => '';
+
   /// Message when the elapsed time is less than 15 seconds.
   @override
   String justNow(int seconds) => 'just ara';