+++ lib/src/messages/languages/ro_msg.dart
@@ -12,6 +12,12 @@ class RomanianMessages implements Messages {
   @override
   String suffixAgo() => '';
 
+  @override
+  String prefixFromNow() => 'peste';
+
+  @override
+  String suffixFromNow() => '';
+
   /// Message when the elapsed time is less than 15 seconds.
   @override
   String justNow(int seconds) => 'tocmai acum';